@@ -15,13 +15,13 @@
         }
     ],
     "require": {
-        "opis/stream": "^1.1",
+        "php": "^7.4",
         "ext-json": "*",
         "ext-fileinfo": "*",
-        "php": "^7.1"
+        "opis/stream": "2020.x-dev"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0"
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "psr-4": {
@@ -40,7 +40,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-master": "2020.x-dev"
         }
     }
 }
