@@ -17,16 +17,16 @@
 
 namespace Opis\FileSystem\Directory;
 
-use Opis\FileSystem\File\IFileInfo;
+use Opis\FileSystem\File\FileInfo;
 use Opis\FileSystem\Handler\CachedHandler;
-use Opis\FileSystem\IProtocolInfo;
+use Opis\FileSystem\ProtocolInfo;
 use Opis\FileSystem\Traits\DirectoryFullPathTrait;
 
-final class CachedDirectory implements IDirectory
+final class CachedDirectory implements Directory
 {
     use DirectoryFullPathTrait;
 
-    /** @var \Opis\FileSystem\Directory\IDirectory */
+    /** @var \Opis\FileSystem\Directory\Directory */
     private $directory;
     /** @var CachedHandler */
     private $handler;
@@ -35,16 +35,16 @@ final class CachedDirectory implements IDirectory
 
     /**
      * CachedDirectory constructor.
-     * @param IDirectory $directory
+     * @param Directory $directory
      * @param CachedHandler $handler
      */
-    public function __construct(IDirectory $directory, CachedHandler $handler)
+    public function __construct(Directory $directory, CachedHandler $handler)
     {
         $this->directory = $directory;
         $this->handler = $handler;
         $this->path = $directory->path();
 
-        if ($directory instanceof IProtocolInfo) {
+        if ($directory instanceof ProtocolInfo) {
             $this->protocol = $directory->protocol();
         }
     }
@@ -60,7 +60,7 @@ final class CachedDirectory implements IDirectory
     /**
      * @inheritDoc
      */
-    public function doNext(): ?IFileInfo
+    public function doNext(): ?FileInfo
     {
         if ($this->directory === null) {
             return null;
