@@ -17,10 +17,10 @@
 
 namespace Opis\FileSystem\Test;
 
-use Opis\FileSystem\File\IFileInfo;
+use Opis\FileSystem\File\FileInfo;
 use Opis\FileSystem\Handler\LocalFileHandler;
-use Opis\FileSystem\IMountManager;
 use Opis\FileSystem\MountManager;
+use Opis\FileSystem\DefaultMountManager;
 use Opis\Stream\PHPDataStream;
 use PHPUnit\Framework\TestCase;
 
@@ -28,7 +28,7 @@ class MountManagerTest extends TestCase
 {
     use FilesTrait;
 
-    /** @var IMountManager */
+    /** @var MountManager */
     protected static $manager;
 
     /** @var string */
@@ -37,11 +37,11 @@ class MountManagerTest extends TestCase
     /**
      * @inheritDoc
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         self::$dir = self::copyFiles(__DIR__ . '/files/manager', __DIR__ . '/files');
 
-        self::$manager = new MountManager([
+        self::$manager = new DefaultMountManager([
             'a' => new LocalFileHandler(self::$dir . '/a'),
             'b' => new LocalFileHandler(self::$dir . '/b'),
         ]);
@@ -50,13 +50,13 @@ class MountManagerTest extends TestCase
     /**
      * @inheritDoc
      */
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
         self::deleteFiles(self::$dir);
     }
 
 
-    protected function m(): IMountManager
+    protected function m(): MountManager
     {
         return self::$manager;
     }
@@ -64,7 +64,7 @@ class MountManagerTest extends TestCase
     public function testCopyFile()
     {
         $info = $this->m()->copy('a://copy-test1.txt', 'b://copy-test-from-a.txt');
-        $this->assertInstanceOf(IFileInfo::class, $info);
+        $this->assertInstanceOf(FileInfo::class, $info);
 
         $this->assertEquals('copy-test-from-a.txt', $info->path());
         $this->assertEquals('b://copy-test-from-a.txt', $info->fullPath());
@@ -73,7 +73,7 @@ class MountManagerTest extends TestCase
     public function testCopyFileSubDir()
     {
         $info = $this->m()->copy('a://copy-test1.txt', 'b://new-dir/copy-test-from-a.txt');
-        $this->assertInstanceOf(IFileInfo::class, $info);
+        $this->assertInstanceOf(FileInfo::class, $info);
 
         $this->assertEquals('copy-test-from-a.txt', $info->name());
         $this->assertEquals('new-dir/copy-test-from-a.txt', $info->path());
@@ -82,11 +82,11 @@ class MountManagerTest extends TestCase
     public function testCopyDir()
     {
         $info = $this->m()->copy('a://copy-dir', 'b://copy-from-a');
-        $this->assertInstanceOf(IFileInfo::class, $info);
+        $this->assertInstanceOf(FileInfo::class, $info);
         $this->assertTrue($info->stat()->isDir());
 
         $info = $this->m()->info('b://copy-from-a/sub-copy-dir/b.txt');
-        $this->assertInstanceOf(IFileInfo::class, $info);
+        $this->assertInstanceOf(FileInfo::class, $info);
         $this->assertTrue($info->stat()->isFile());
 
     }
@@ -94,7 +94,7 @@ class MountManagerTest extends TestCase
     public function testMoveFile()
     {
         $info = $this->m()->rename('a://move-file1.txt', 'b://move-sub-dir/move-sub-dir2/moved-from-a.txt');
-        $this->assertInstanceOf(IFileInfo::class, $info);
+        $this->assertInstanceOf(FileInfo::class, $info);
 
         $this->assertNull($this->m()->info('a://move-file1.txt'));
     }
@@ -102,11 +102,11 @@ class MountManagerTest extends TestCase
     public function testMoveDir()
     {
         $info = $this->m()->rename('a://move-dir', 'b://moved-from-a');
-        $this->assertInstanceOf(IFileInfo::class, $info);
+        $this->assertInstanceOf(FileInfo::class, $info);
         $this->assertTrue($info->stat()->isDir());
 
         $info = $this->m()->info('b://moved-from-a/sub-move-dir/b.txt');
-        $this->assertInstanceOf(IFileInfo::class, $info);
+        $this->assertInstanceOf(FileInfo::class, $info);
         $this->assertTrue($info->stat()->isFile());
 
         $this->assertNull($this->m()->info('a://move-dir'));
